tools/experiments/PDF_to_DOCX/PDF_to_DOCX/SeekHub_Demo-main/translator/requirements.txt
fastapi
uvicorn[standard]
google-genai>=1.7.0
tiktoken
